chore: refactor SSR package
