+++ packages/ssr/package.json
@@ -1,6 +1,7 @@
 {
   "name": "@times-components/ssr",
-  "private": true,
+  "main": "dist/ssr/server",
+  "dev": "src/ssr/server",
   "version": "1.3.5",
   "scripts": {
     "cleanup-dist": "rm -rf dist",
