+++ packages/ssr/src/client/article.js
@@ -1,0 +1,6 @@
+/* eslint-env browser */
+
+const article = require("../component/article");
+const runClient = require("../lib/run-client");
+
+runClient(article, window.nuk.id, window.nuk.adConfig);
