+++ packages/ssr/src/client/author-profile.js
@@ -1,0 +1,6 @@
+/* eslint-env browser */
+
+const authorProfile = require("../component/author-profile");
+const runClient = require("../lib/run-client");
+
+runClient(authorProfile, window.nuk.slug, window.nuk.page);
