+++ packages/ssr/src/component/topic.js
@@ -2,11 +2,11 @@
 
 const React = require("react");
 const { ApolloProvider } = require("react-apollo");
-const { AuthorProfileProvider } = require("@times-components/provider/rnw");
+const { TopicProvider } = require("@times-components/provider/rnw");
 const Context = require("@times-components/context/rnw").default;
 const { scales } = require("@times-components/styleguide/rnw");
-const AuthorProfile = require("@times-components/author-profile/rnw").default;
-const makeArticleUrl = require("./make-url");
+const Topic = require("@times-components/topic/rnw").default;
+const makeArticleUrl = require("../lib/make-url");
 
 const scale = scales.large;
 const sectionColour = "#FFFFFF";
@@ -16,26 +16,26 @@ module.exports = (client, slug, page) =>
     ApolloProvider,
     { client },
     React.createElement(
-      AuthorProfileProvider,
+      TopicProvider,
       {
         debounceTimeMs: 0,
         slug
       },
-      ({ author, isLoading, error, refetch }) =>
+      ({ isLoading, error, refetch, topic }) =>
         React.createElement(
           Context.Provider,
           { value: { makeArticleUrl, theme: { scale, sectionColour } } },
-          React.createElement(AuthorProfile, {
+          React.createElement(Topic, {
             adConfig: {},
             analyticsStream: () => {},
-            author,
             error,
             isLoading,
             onArticlePress: () => {},
             onTwitterLinkPress: () => {},
             page,
             refetch,
-            slug
+            slug,
+            topic
           })
         )
     )
