+++ packages/ssr/src/server/author-profile.js
@@ -1,0 +1,4 @@
+const authorProfile = require("../component/author-profile");
+const runServer = require("../lib/run-server");
+
+module.exports = (slug, page) => runServer(authorProfile, slug, page);
