+++ packages/ssr/src/server/topic.js
@@ -1,0 +1,4 @@
+const topic = require("../component/topic");
+const runServer = require("../lib/run-server");
+
+module.exports = (slug, page) => runServer(topic, slug, page);
