--- packages/ssr/topic.client.js
@@ -1,7 +1,0 @@
-/* eslint-env browser */
-
-const topic = require("./topic");
-const apolloClient = require("./make-client-ac");
-const runClient = require("./run-client");
-
-runClient(topic(apolloClient(), window.nuk.slug, window.nuk.page));
