+++ packages/ssr/webpack.config.prod.js
@@ -1,37 +1,8 @@
-const path = require("path");
-const webpack = require("webpack");
+const webpackConfigCommon = require("./webpack.config.common");
 
-const alias = { "react-native$": "react-native-web" };
-const extensions = [".web.js", ".js"];
-const mode = "production";
-
-const babelConfig = {
-  test: /\.js$/,
-  use: {
-    loader: "babel-loader",
-    options: {
-      cacheDirectory: true,
-      plugins: ["add-react-displayname", "styled-components"],
-      presets: ["react-native"]
-    }
-  }
-};
-
-if (!process.env.GRAPHQL_ENDPOINT) {
-  throw new Error("You need to set a GRAPHQL_ENDPOINT envar");
-}
-
-module.exports = {
+module.exports = webpackConfigCommon({
   devtool: "source-map",
-  entry: {
-    article: "./article.client.js",
-    "author-profile": "./author-profile.client.js",
-    topic: "./topic.client.js"
-  },
-  mode,
-  module: {
-    rules: [babelConfig]
-  },
+  mode: "production",
   optimization: {
     occurrenceOrder: true,
     splitChunks: {
@@ -44,20 +15,5 @@ module.exports = {
         }
       }
     }
-  },
-  output: {
-    filename: "[name].bundle.js",
-    path: path.resolve(__dirname, "./dist")
-  },
-  plugins: [
-    new webpack.DefinePlugin({
-      "process.env": {
-        GRAPHQL_ENDPOINT: JSON.stringify(process.env.GRAPHQL_ENDPOINT)
-      }
-    })
-  ],
-  resolve: {
-    alias,
-    extensions
   }
-};
+});
